Show that inverse charts are C^\infty also.
The idea is the same as for normal charts; the details are harder.
Normal charts (in my case) are defined on the whole domain;
inverses restrict to an open set. That means showing membership
of an atlas is harder, and in smoothness an extra term occurs
which we have to rewrite away.